hash: Fix invalid shift in hash table code

+++ src/lib/util/hash.c
@@ -150,10 +150,17 @@ static uint8_t parent_byte[256] = {
  */
 static uint32_t reverse(uint32_t key)
 {
-	return ((reversed_byte[key & 0xff] << 24) |
-		(reversed_byte[(key >> 8) & 0xff] << 16) |
-		(reversed_byte[(key >> 16) & 0xff] << 8) |
-		(reversed_byte[(key >> 24) & 0xff]));
+	/*
+	 *	Cast to uint32_t is required because the
+	 *	default type of of the expression is an
+	 *	int and ubsan correctly complains that
+	 *	the result of 0xff << 24 won't fit in an
+	 *	unsigned 32bit integer.
+	 */
+	return (((uint32_t)reversed_byte[key & 0xff] << 24) |
+		((uint32_t)reversed_byte[(key >> 8) & 0xff] << 16) |
+		((uint32_t)reversed_byte[(key >> 16) & 0xff] << 8) |
+		((uint32_t)reversed_byte[(key >> 24) & 0xff]));
 }
 
 /*
